Organize GUI Package
Added the Helper GUIs to the viewHelper package to clean up the file system.
