+++ src/view/viewHelper/MazeGUI.java
@@ -1,4 +1,4 @@
-package view;
+package view.viewHelper;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
