+++ src/view/viewHelper/QuestionRoomGUI.java
@@ -1,4 +1,4 @@
-package view;
+package view.viewHelper;
 
 import java.awt.Color;
 import java.awt.Dimension;
